t.float in orders table

+++ db/migrate/20210119214321_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[6.0]
   def change
     create_table :orders do |t|
       t.references :user, null: false, foreign_key: true
-      t.integer :total_price
+      t.float :total_price
       t.datetime :date
 
       t.timestamps
